% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_covariate_data.R
\name{attach_covariate_data}
\alias{attach_covariate_data}
\title{Attach covariate data}
\usage{
attach_covariate_data(
  se,
  covariate,
  dataset = NULL,
  columns = NULL,
  date_col = "sample_date"
)
}
\arguments{
\item{se}{Sample events from \code{mermaidr}}

\item{covariate}{Covariate to attach. Both covariate title or ID are permitted. Run \code{list_covariates()} to see available covariates.}

\item{dataset}{Dataset within the covariate. Not required when there is only one dataset. Run \code{\link{list_datasets_for_covariate}} to see datasets and columns.}

\item{columns}{Columns within the dataset. When NULL (the default), returns all columns. Run \code{\link{list_datasets_for_covariate}} to see datasets and bands.}
}
\description{
Attach covariate data. Not date dependent -- only attach based on sample event location.
}
