% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_helper.R
\name{covariate_helper}
\alias{covariate_helper}
\title{Covariate helper}
\usage{
covariate_helper(covariate)
}
\arguments{
\item{covariate}{Covariate name or ID}
}
\description{
Tells you what function to use to get covariates data! Also returns information on datasets, bands or columns, and more.
}
\examples{
covariate_helper("Daily Sea Surface Temperature")
covariate_helper("GPW MEOW Realms")
}
