% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covariate_type.R
\name{get_covariate_type}
\alias{get_covariate_type}
\title{Get covariate type}
\usage{
get_covariate_type(covariate)
}
\arguments{
\item{covariate}{Covariate name or ID}
}
\description{
Tells you covariate type (raster, vector, raster and vectors, or unknown)
}
\examples{
# get_covariate_type("Daily Global 5km Satellite Sea Surface Temperature (CoralTemp)")
# get_covariate_type("GPW MEOW Realms")
}
