% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets_for_covariate.R
\name{list_datasets_for_covariate}
\alias{list_datasets_for_covariate}
\title{List datasets and bands or columns for a covariate}
\usage{
list_datasets_for_covariate(covariate)
}
\arguments{
\item{covariate}{Covariate}
}
\description{
List datasets and bands or columns for a covariate
}
