% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_covariates.R
\name{search_covariates}
\alias{search_covariates}
\title{Search available covariates}
\usage{
search_covariates(title, description, as_data_frame = TRUE)
}
\arguments{
\item{title}{Search term within title. Case insensitive.}

\item{description}{Search term within description. Case insensitive.}

\item{as_data_frame}{Whether to return the covariates as a data frame. Defaults to \code{TRUE}.}
}
\description{
Search available covariates
}
\examples{
# search_covariates("MEOW")
}
