% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_zonal_statistics.R
\name{summarise_zonal_statistics}
\alias{summarise_zonal_statistics}
\title{Summarise zonal statistics}
\usage{
summarise_zonal_statistics(
  zonal_statistics,
  temporal_stats = c("min", "max", "mean")
)
}
\arguments{
\item{zonal_statistics}{Sample events with zonal statistics, from \code{get_zonal_statistics}.}

\item{temporal_stats}{Temporal statistics -- used to summarise the data over time.}
}
\description{
Summarise zonal statistics. Using the results from  \code{get_zonal_statistics},
this summarises the data for *all `n_days`* prior to the sample event date, using
`temporal_stats` to summarise the resulting data over time.
}
