% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_stats.R
\name{summary_zonal_stats}
\alias{summary_zonal_stats}
\title{Get summary of zonal statistics}
\usage{
summary_zonal_stats(
  df,
  collection,
  n_days = 365,
  buffer = 1000,
  stats = c("min", "max", "mean")
)
}
\arguments{
\item{df}{Sample events from \code{mermaidr}}

\item{collection}{Covariates to get statistics for}

\item{n_days}{Number of days prior to sample date to get statistics for. Defaults to 365.}

\item{buffer}{Buffer around site location, in metres. Defaults to 1000.}

\item{stats}{Summary statistics. One of: min, max, or mean.}
}
\description{
Get summary of zonal statistics
}
\examples{
# TODO
}
