% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonal_stats.R
\name{get_zonal_statistics}
\alias{get_zonal_statistics}
\title{Get summary of zonal statistics}
\usage{
get_zonal_statistics(
  se,
  covariate,
  n_days = 365,
  buffer = 1000,
  stats = c("min", "max", "mean")
)
}
\arguments{
\item{se}{Sample events from \code{mermaidr}}

\item{covariate}{Covariate to get statistics for. Both covariate title or ID are permitted.}

\item{n_days}{Number of days to get statistics for. Includes the sample date
itself, and days prior to it -- e.g., 365 days would include the sample date
and the 364 days prior. Defaults to 365.}

\item{buffer}{Buffer around site location, in metres. Defaults to 1000.}

\item{stats}{Summary statistics. One of: min, max, or mean.}
}
\description{
Get summary of zonal statistics
}
